// frontend/src/api/talos/perf.pb.ts
// This file is auto-generated by internal/internal/tools/talostsgen. DO NOT EDIT.
//
// Source of truth: yaml struct tags on the Go types in
// github.com/siderolabs/talos/pkg/machinery/resources/

export type CPUStat = {
  user?: number
  nice?: number
  system?: number
  idle?: number
  iowait?: number
  irq?: number
  softIrq?: number
  steal?: number
  guest?: number
  guestNice?: number
}

export type CPUSpec = {
  cpu?: CPUStat[]
  cpuTotal?: CPUStat
  irqTotal?: number
  contextSwitches?: number
  processCreated?: number
  processRunning?: number
  processBlocked?: number
  softIrqTotal?: number
}

export type MemorySpec = {
  total?: number
  used?: number
  available?: number
  buffers?: number
  cached?: number
  swapCached?: number
  active?: number
  inactive?: number
  activeAnon?: number
  inactiveAnon?: number
  activeFile?: number
  inactiveFile?: number
  unevictable?: number
  mlocked?: number
  swapTotal?: number
  swapFree?: number
  dirty?: number
  writeback?: number
  anonPages?: number
  mapped?: number
  shmem?: number
  slab?: number
  sreclaimable?: number
  sunreclaim?: number
  kernelStack?: number
  pageTables?: number
  nfsunstable?: number
  bounce?: number
  writeBacktmp?: number
  commitLimit?: number
  commitTedas?: number
  vmallocTotal?: number
  vmallocUsed?: number
  vmallocChunk?: number
  hardwareCorrupted?: number
  anonHugePages?: number
  shmemHugePages?: number
  shmemPmdMapped?: number
  cmaTotal?: number
  cmaFree?: number
  hugePagesTotal?: number
  hugePagesFree?: number
  hugePagesRsvd?: number
  hugePagesSurp?: number
  hugepagesize?: number
  directMap4k?: number
  directMap2m?: number
  directMap1g?: number
}